+++ CHANGES.md
@@ -10,6 +10,12 @@ To be released.
 
 ### Backward-incompatible API changes
 
+ -  The signature of `Address(ImmutableArray<byte>)` constructor became
+    `Address(in ImmutableArray<byte>)`.  [[#2756]]
+ -  The signature of `BlockHash(ImmutableArray<byte>)` constructor became
+    `BlockHash(in ImmutableArray<byte>)`.  [[#2756]]
+ -  The signature of `Nonce(ImmutableArray<byte>)` constructor became
+    `Nonce(in ImmutableArray<byte>)`.  [[#2756]]
  -  Removed `IAccountStateView.GetValidatorSet()` method.  [[#2733]]
  -  Removed `IAccountStateDelta.SetValidator(Validator)` method.  [[#2733]]
  -  (Libplanet.Extensions.Cocona)  Dropped .NET Standard 2.0 and .NET Core 3.1
@@ -29,6 +35,8 @@ To be released.
 
 ### Added APIs
 
+ -  Added `ByteUtil.ParseHexToImmutable()` static method.  [[#2756]]
+
 ### Behavioral changes
 
 ### Bug fixes
@@ -40,6 +48,7 @@ To be released.
 [#2732]: https://github.com/planetarium/libplanet/pull/2732
 [#2733]: https://github.com/planetarium/libplanet/pull/2733
 [#2743]: https://github.com/planetarium/libplanet/pull/2743
+[#2756]: https://github.com/planetarium/libplanet/pull/2756
 
 
 Version 0.46.0